- util/pyobject_call.tmpl.inl: help compiler to pass on (const) reference arguments. - util/py_stl.h: make floats, ints and longs acceptable as complex numbers - num/filters: add static cast to avoid warning - prim::Aabb3D: add vectorial grow - test/test.cpp: allow to set filename of logfile on commandline
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@759 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/num/filters.inl
@@ -48,7 +48,7 @@ namespace impl
 		TPolynomial oneMinZ = 2 * samplingFrequency * (TNumTraits::one - TPolynomial::x());
 		TPolynomial onePlusZ = TNumTraits::one + TPolynomial::x();
 		TPolynomial result;
-		const int n = std::distance(first, last);
+		const int n = static_cast<int>(std::distance(first, last));
 		int i = 0;
 		while (first != last)
 		{
